services can be deleted

+++ app/Http/Livewire/Provider/ServiceProviderDashboard.php
@@ -3,6 +3,8 @@
 namespace App\Http\Livewire\Provider;
 
 use App\Models\Service;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -18,12 +20,12 @@ class ServiceProviderDashboard extends Component
     public function render()
     {
         $services = Service::whereuser_id(auth()->id())
-            ->wherestatus(1)
-            ->latest()->paginate(12);
+        ->wherestatus(1)
+        ->latest()->paginate(12);
 
         $inactive_services = Service::whereuser_id(auth()->id())
-            ->wherestatus(0)
-            ->latest()->paginate(12);
+        ->wherestatus(0)
+        ->latest()->paginate(12);
 
         return view('livewire.provider.service-provider-dashboard', compact('services', 'inactive_services'))->extends('layouts.master');
     }
@@ -73,15 +75,33 @@ class ServiceProviderDashboard extends Component
 
     public function deleteService(Service $service)
     {
-        $service->providerServices()->delete();
-        $service->delete();
-        $this->dispatchBrowserEvent('swal:modal', [
-            'type' => 'success',
-            'title' => $service->name .' was deleted successfully!',
-            'text' => '',
-        ]);
-        $this->myActiveServices();
+        DB::beginTransaction();
+        if($service) {
+            $photos = explode('|', $service->images);
+            $originalPath = 'uploads/services/images/';
+            // Remove old images
+            if(count($photos) > 0){
+                //  dd($photos);
+                foreach($photos as $photo){
+                    // Delete old image from file
+                    Storage::disk('s3')->delete(parse_url($originalPath .$photo));
+                }
+                $service->providerServices()->delete();
+                $service->delete();
+                DB::commit();
+                $this->dispatchBrowserEvent('swal:modal', [
+                    'type' => 'success',
+                    'title' => $service->name .' was deleted successfully!',
+                    'text' => '',
+                ]);
+                $this->myActiveServices();
+            }
+            }else{
+                DB::rollBack();
+                return redirect()->back();
+            }
+
+        }
+
+
     }
-
-
-}
